refactor: clean up components file structure
